#11 unkown term in query
The script now deals with single-word queries where the term is unknown and multiple-word queries when the query is not "NOT unknownword".

I am not sure how to get this aforementioned functionality working. To deal with unknown terms, the script matches an UKN key in the dictionary which works, but if the user queries "NOT unknownword", the output is incorrect. This is because there's a false positive match found in the last document (matching to the UKN key).