Fixed the erroneous printing of an empty document
The script now produces the correct output for searching "NOT Angola" = 92 matching documents (instead of 93 like previously). My fix is not very sophisticated as I was not able to fix the underlying issue, but using conditionals, it is easy to remove the empty match from the end of the output using 'brute force'.